+++ sdk/messaging/azservicebus/liveTestHelpers_test.go
@@ -49,10 +49,12 @@ func createQueue(t *testing.T, connectionString string, queueProperties *admin.Q
 		queueProperties = &admin.QueueProperties{}
 	}
 
-	autoDeleteOnIdle := 5 * time.Minute
+	autoDeleteOnIdle := "PT5M"
 	queueProperties.AutoDeleteOnIdle = &autoDeleteOnIdle
 
-	_, err = adminClient.CreateQueue(context.Background(), queueName, queueProperties, nil)
+	_, err = adminClient.CreateQueue(context.Background(), queueName, &admin.CreateQueueOptions{
+		Properties: queueProperties,
+	})
 	require.NoError(t, err)
 
 	return queueName, func() {